cswrap: use remove_self_from_path() to avoid a deadlock
... while scanning atlas-3.10.3-1.fc27 with cswrap/csclng.  The package
contains a configure check that executes 'gcc -v' using an absolute path
to the cswrap symlink.  This transitively executes cswrap again through
csclng.  Consequently, if both cswrap instances try to write something
to the capture file, they end up waiting for each other indefinitely to
unlock the semaphore.

The remove_self_from_path() fnc was taken from cscppc and put to the
cswrap-util module so that it can be shared by both the projects and
maintained at a single place.