csexec: dynamic linker wrapper (x86_64 only for now)
This is a proof-of-concept execution wrapper.  The ultimate goal is to
run dynamic analyzers and formal verifiers on RPM packages fully
automatically.  We already know how to tweak the build process, how to
hook static analyzers on compiler invocation, and how to capture their
results.  This is, however, not so easy with dynamic analyzers.

The %check section of RPM packages can use arbitrary testing frameworks
and scripting languages to verify the binaries produced in the %build
and %install sections.  Our goal is to dynamically analyze (or formally
verify) the produced binaries.  At the same time, we want to interfere
with all the testing frameworks as little as possible.

This wrapper runs the dynamic linker explicitly.  Optionally, it inserts
an execution wrapper if the CSEXEC_WRAP_CMD environment variable is set
at run time.  The string may optionally contain BEL-separated list of
arguments that will be passed to the wrapper in front of the path to the
dynamic linker exectuable.

Note that the current implementation will only work on the x86_64 arch.

Closes: https://github.com/kdudka/cswrap/pull/2