csexec-loader: custom ELF interpreter to load csexec
This binary can be set as ELF interpreter while linking binaries in
%build and %install sections of RPM packages.  The wrapper executes
/usr/bin/csexec passing it EXECFN as argv[0], followed by original
command-line arguments.  Then csexec can optionally run the real
wrapper if ${CSEXEC_WRAP_CMD} is set.  Otherwise csexec should take
care of executing the original ELF file natively.

There are some constrains on the implementation of ELF interpreters.
Namely it does not work well if an ELF interpreter itself uses the
original ELF interpreter or C run-time in general, which is tightly
coupled with the system ELF interpreter.  Therefore this wrapper is
implemented such that it does not rely on any C run-time libraries.

Note that the current implementation will only work on the x86_64 arch.

Closes: https://github.com/kdudka/cswrap/pull/2